rust/pv_*: Add more deny lints to pv and pv_core
Denies compiling if one of the following lints find something in pv or
pv_core:
    missing_docs,
    missing_debug_implementations,
    trivial_numeric_casts,
    unstable_features,
    unused_import_braces,
    unused_qualifications

Those lint force developers to avoid unnecessary code and providing
debuggability & documentation for each public symbol.

Fix the compile time error introduced with those lints.

+++ rust/pv/src/openssl_extensions/stackable_crl.rs
@@ -111,7 +111,7 @@ impl StackableX509Crl {
                 );
                 if r.is_null() {
                     let err = openssl_sys::ERR_peek_last_error();
-                    if openssl_sys::ERR_GET_LIB(err) as c_int == openssl_sys::ERR_LIB_PEM
+                    if openssl_sys::ERR_GET_LIB(err) == openssl_sys::ERR_LIB_PEM
                         && openssl_sys::ERR_GET_REASON(err) == openssl_sys::PEM_R_NO_START_LINE
                     {
                         openssl_sys::ERR_clear_error();
